+++ Loginator/Converter/ObjectToBoolConverter.cs
@@ -1,16 +1,19 @@
-﻿using System;
+﻿// Copyright (C) 2024 Claudia Wagner, Daniel Kuster
+
+using System;
 using System.Globalization;
 using System.Windows.Data;
 
 namespace Loginator.Converter {
-    public class IsEqualsConverter : IValueConverter
-    {
+
+    public class ObjectToBoolConverter : IValueConverter {
+
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture) {
             return value.Equals(parameter);
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture) {
-            throw new NotImplementedException();
+            return value.Equals(true) ? parameter : Binding.DoNothing;
         }
     }
 }