+++ Loginator/Views/LogDetailsView.xaml.cs
@@ -1,0 +1,14 @@
+﻿using System.Windows.Controls;
+
+namespace Loginator.Views {
+
+    /// <summary>
+    /// Interaction logic for LogDetailsView.xaml
+    /// </summary>
+    public partial class LogDetailsView : UserControl {
+
+        public LogDetailsView() {
+            InitializeComponent();
+        }
+    }
+}