+++ Loginator/Views/NamespacesView.xaml.cs
@@ -1,0 +1,14 @@
+﻿using System.Windows.Controls;
+
+namespace Loginator.Views {
+
+    /// <summary>
+    /// Interaction logic for NamespacesView.xaml
+    /// </summary>
+    public partial class NamespacesView : UserControl {
+
+        public NamespacesView() {
+            InitializeComponent();
+        }
+    }
+}